Finalização do CRUD de Alunos com Modal

+++ src/main/java/com/apisys/teste/repository/AlunoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.apisys.teste.domain.Aluno;
 
 @Repository
-public interface AlunoRepository extends JpaRepository<Aluno, Integer>{
+public interface AlunoRepository extends JpaRepository<Aluno, Long>{
 
 }
