+++ src/main/java/com/apisys/teste/repository/CursoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.apisys.teste.domain.Curso;
 
 @Repository
-public interface CursoRepository extends JpaRepository<Curso, Integer>{
+public interface CursoRepository extends JpaRepository<Curso, Long>{
 
 }
